# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Time Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: Client found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /clients/{clientId}/assignments:
    get:
      operationId: listClientAssignments
      tags: [clients]
      summary: List users assigned to a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: Assigned users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: assignUserToClient
      tags: [clients]
      summary: Assign a user to a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAssignmentInput"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /clients/{clientId}/assignments/{userId}:
    delete:
      operationId: removeUserFromClient
      tags: [clients]
      summary: Remove a user from a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /clients/{clientId}/requesters:
    get:
      operationId: listClientRequesters
      tags: [clients]
      summary: List the client-side people who request work
      parameters:
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: Requesters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientRequester"
    post:
      operationId: addClientRequester
      tags: [clients]
      summary: Record a new requester on the client's side
      parameters:
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientRequesterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRequester"
  /clients/{clientId}/requesters/{requesterId}:
    delete:
      operationId: removeClientRequester
      tags: [clients]
      summary: Remove a requester from a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
        - name: requesterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /clients/{clientId}/fte-history:
    get:
      operationId: listClientFteHistory
      tags: [clients]
      summary: List FTE history for a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: FTE history entries ordered by effectiveFrom
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientFteHistory"
    post:
      operationId: addClientFteHistory
      tags: [clients]
      summary: Add an FTE history entry for a client
      parameters:
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientFteHistoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientFteHistory"
  /clients/{clientId}/fte-history/{entryId}:
    delete:
      operationId: deleteClientFteHistory
      tags: [clients]
      summary: Delete an FTE history entry
      parameters:
        - $ref: "#/components/parameters/ClientId"
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Projects ───────────────────────────────────────────────────────────
...
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── Auth ────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [analyst, associate, avp, md]
        title:
          description: >-
            Overrides the role's default label, e.g. "Admin" for someone who
            holds the md permission rank because they administer the system
            rather than because they are the Managing Director. Null shows the
            ordinary role label.
          type: ["string", "null"]
        reportingToId:
          type: ["integer", "null"]
        reportingToName:
          type: ["string", "null"]
        tourCompletedAt:
          description: >-
            When this person finished or dismissed the guided tour. Null means
            they have never been shown it, which is what opens it by itself on
            a first sign-in.
          type: ["string", "null"]
          format: date-time
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
...
    Client:
      type: object
      required: [id, name, fteCount, engagementType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        fteCount:
          type: number
          minimum: 0.1
          maximum: 100
          description: "Number of full-time equivalents; 1 FTE = 8h/day, 40h/week, 160h/month"
        engagementType:
          $ref: "#/components/schemas/EngagementType"
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
    ClientFteHistory:
      type: object
      required: [id, clientId, fteCount, effectiveFrom, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        fteCount:
          type: number
          minimum: 0.1
          maximum: 100
        effectiveFrom:
          type: string
          format: date
          description: "First day this FTE count applies (YYYY-MM-DD)"
        effectiveTo:
          type: ["string", "null"]
          format: date
          description: "Last day this FTE count applies; null means open-ended (current)"
        createdAt:
          type: string
          format: date-time
    ClientFteHistoryInput:
      type: object
      required: [fteCount, effectiveFrom]
      properties:
        fteCount:
          type: number
          minimum: 0.1
          maximum: 100
        effectiveFrom:
          type: string
          format: date
        effectiveTo:
          type: ["string", "null"]
          format: date
    ClientInput:
      type: object
      # A client with no associate on it is visible to nobody below MD, since
      # every rank beneath reaches clients through assignment.
      required: [name, description, associateIds]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        fteCount:
          type: number
          minimum: 0.1
          maximum: 100
          default: 1
        engagementType:
          $ref: "#/components/schemas/EngagementType"
        associateIds:
          type: array
          minItems: 1
          items:
            type: integer
          description: "IDs of Associates to assign as responsible"
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        fteCount:
          type: number
          minimum: 0.1
          maximum: 100
        engagementType:
          $ref: "#/components/schemas/EngagementType"
        isActive:
          type: boolean

    # ─── Projects ─────────────────────────────────────────────────────────
...
    EngagementType:
      type: string
      enum: [fte, block_hours, product]
      description: "How a client is engaged: dedicated FTEs, a purchased block of hours, or defined deliverables"
...
    UserAssignmentInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
...
    ClientRequester:
      type: object
      required: [id, clientId, name, designation, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        name:
          type: string
        designation:
          type: string
          description: Their title at the client, e.g. "CFO".
        createdAt:
          type: string
          format: date-time
    ClientRequesterInput:
      type: object
      required: [name, designation]
      properties:
        name:
          type: string
        designation:
          type: string
...